Moving some tests to test_adapters.py
