+++ python/cuml/test/test_adapters.py
@@ -14,12 +14,23 @@
 # limitations under the License.
 #
 
-import cupy as cp
 import pytest
 
+import cupy as cp
+import numpy as np
 from cupyx.scipy.sparse import coo_matrix
 
 from cuml.thirdparty_adapters.adapters import check_array
+from cuml.thirdparty_adapters.sparsefuncs_fast import \
+    csr_mean_variance_axis0, \
+    csc_mean_variance_axis0, \
+    _csc_mean_variance_axis0, \
+    inplace_csr_row_normalize_l1, \
+    inplace_csr_row_normalize_l2
+
+from cuml.test.test_preproc_utils import sparse_clf_dataset  # noqa: F401
+from cuml.test.test_preproc_utils import assert_allclose
+from sklearn.preprocessing import normalize as sk_normalize
 
 
 def test_check_array():
@@ -67,3 +78,96 @@ def test_check_array():
     check_array(arr, ensure_min_features=0)
     with pytest.raises(ValueError):
         check_array(arr, ensure_min_features=1)
+
+
+def test_csr_mean_variance_axis0(sparse_clf_dataset):  # noqa: F811
+    X_np, X = sparse_clf_dataset
+
+    if not cp.sparse.issparse(X):
+        pytest.skip("Skipping non-CuPy or non-sparse arrays")
+
+    if X.format != 'csr':
+        X = X.tocsr()
+
+    means, variances = csr_mean_variance_axis0(X)
+
+    X_np = X_np.toarray()
+    ref_means = np.nanmean(X_np, axis=0)
+    ref_variances = np.nanvar(X_np, axis=0)
+
+    assert_allclose(means, ref_means)
+    assert_allclose(variances, ref_variances)
+
+
+def test_csc_mean_variance_axis0(sparse_clf_dataset):  # noqa: F811
+    X_np, X = sparse_clf_dataset
+
+    if not cp.sparse.issparse(X):
+        pytest.skip("Skipping non-CuPy or non-sparse arrays")
+
+    if X.format != 'csc':
+        X = X.tocsc()
+
+    means, variances = csc_mean_variance_axis0(X)
+
+    X_np = X_np.toarray()
+    ref_means = np.nanmean(X_np, axis=0)
+    ref_variances = np.nanvar(X_np, axis=0)
+
+    assert_allclose(means, ref_means)
+    assert_allclose(variances, ref_variances)
+
+
+def test__csc_mean_variance_axis0(sparse_clf_dataset):  # noqa: F811
+    X_np, X = sparse_clf_dataset
+
+    if not cp.sparse.issparse(X):
+        pytest.skip("Skipping non-CuPy or non-sparse arrays")
+
+    if X.format != 'csc':
+        X = X.tocsc()
+
+    means, variances, counts_nan = _csc_mean_variance_axis0(X)
+
+    X_np = X_np.toarray()
+    ref_means = np.nanmean(X_np, axis=0)
+    ref_variances = np.nanvar(X_np, axis=0)
+    ref_counts_nan = np.isnan(X_np).sum(axis=0)
+
+    assert_allclose(means, ref_means)
+    assert_allclose(variances, ref_variances)
+    assert_allclose(counts_nan, ref_counts_nan)
+
+
+def test_inplace_csr_row_normalize_l1(sparse_clf_dataset):  # noqa: F811
+    X_np, X = sparse_clf_dataset
+
+    if not cp.sparse.issparse(X):
+        pytest.skip("Skipping non-CuPy or non-sparse arrays")
+
+    if X.format != 'csr':
+        X = X.tocsr()
+
+    inplace_csr_row_normalize_l1(X)
+
+    X_np = X_np.toarray()
+    X_np = sk_normalize(X_np, norm='l1', axis=1)
+
+    assert_allclose(X, X_np)
+
+
+def test_inplace_csr_row_normalize_l2(sparse_clf_dataset):  # noqa: F811
+    X_np, X = sparse_clf_dataset
+
+    if not cp.sparse.issparse(X):
+        pytest.skip("Skipping non-CuPy or non-sparse arrays")
+
+    if X.format != 'csr':
+        X = X.tocsr()
+
+    inplace_csr_row_normalize_l2(X)
+
+    X_np = X_np.toarray()
+    X_np = sk_normalize(X_np, norm='l2', axis=1)
+
+    assert_allclose(X, X_np)
